Price to float

+++ geekshop/orders/views.py
@@ -122,7 +122,7 @@ class OrderDetail(DetailView, BaseClassContextMixin):
 
     @staticmethod
     def get_product_price(request, product_id):
-        price = int()
+        price = float()
         try:
             price = Product.objects.filter(id=product_id).first().price
         except AttributeError as e:
